# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/authors/search": {
      "get": {
        "operationId": "AuthorController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/check-username": {
      "get": {
        "description": "Checks if a username is available for use. Returns true if available or if it belongs to the current user.",
        "operationId": "AuthorController_checkUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "description": "Username to check",
            "schema": {
              "example": "johndoe",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Username availability status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check username availability",
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors": {
      "get": {
        "description": "Retrieves a paginated list of all authors.",
        "operationId": "AuthorController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of authors retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "author_123"
                      },
                      "username": {
                        "type": "string",
                        "example": "johndoe"
                      },
                      "display_name": {
                        "type": "string",
                        "example": "John Doe"
                      },
                      "email": {
                        "type": "string",
                        "example": "john@example.com"
                      },
                      "bio": {
                        "type": "string",
                        "example": "Software developer and writer"
                      },
                      "website": {
                        "type": "string",
                        "example": "https://johndoe.com"
                      },
                      "avatar_url": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ]
      },
      "post": {
        "description": "Creates a new author profile. Requires authentication.",
        "operationId": "AuthorController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The author has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "author_123"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "display_name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@example.com"
                    },
                    "bio": {
                      "type": "string",
                      "example": "Software developer and writer"
                    },
                    "website": {
                      "type": "string",
                      "example": "https://johndoe.com"
                    },
                    "avatar_url": {
                      "type": "string",
                      "example": "https://example.com/avatar.jpg"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new author",
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}": {
      "get": {
        "description": "Retrieves a specific author by their UUID or username. Prefix identifier with @ to search by username (e.g., @johndoe), otherwise looks up by UUID.",
        "operationId": "AuthorController_getByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "description": "Author UUID (e.g., \"550e8400-e29b-41d4-a716-446655440000\") or @username (e.g., \"@johndoe\")",
            "schema": {
              "example": "@johndoe",
              "type": "string"
            }
          },
          {
            "name": "embed",
            "required": false,
            "in": "query",
            "description": "Embed related resources. Use \"home\" to include the home crux.",
            "schema": {
              "enum": [
                "home"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The author data",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "description": "Author without embedded resources",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "author_123"
                        },
                        "username": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "displayName": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "bio": {
                          "type": "string",
                          "example": "Software developer and writer"
                        },
                        "accountId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "550e8400-e29b-41d4-a716-446655440000"
                        },
                        "homeId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "550e8400-e29b-41d4-a716-446655440001"
                        },
                        "type": {
                          "type": "string",
                          "example": "individual"
                        },
                        "kind": {
                          "type": "string",
                          "example": "writer"
                        },
                        "meta": {
                          "type": "object",
                          "example": {}
                        },
                        "created": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "description": "Author with embedded home crux (when ?embed=home)",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "author_123"
                        },
                        "username": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "displayName": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "bio": {
                          "type": "string",
                          "example": "Software developer and writer"
                        },
                        "accountId": {
                          "type": "string",
                          "example": "account_456"
                        },
                        "type": {
                          "type": "string",
                          "example": "individual"
                        },
                        "kind": {
                          "type": "string",
                          "example": "writer"
                        },
                        "meta": {
                          "type": "object",
                          "example": {}
                        },
                        "home": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "crux_789"
                            },
                            "title": {
                              "type": "string",
                              "example": "My Home Crux"
                            }
                          }
                        },
                        "created": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Author not found"
          }
        },
        "summary": "Get an author by ID or username",
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{id}": {
      "patch": {
        "description": "Updates an existing author profile. Requires authentication and ownership.",
        "operationId": "AuthorController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the author to update",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The author has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "author_123"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "display_name": {
                      "type": "string",
                      "example": "John Updated"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.updated@example.com"
                    },
                    "bio": {
                      "type": "string",
                      "example": "Updated bio information"
                    },
                    "website": {
                      "type": "string",
                      "example": "https://johnupdated.com"
                    },
                    "avatar_url": {
                      "type": "string",
                      "example": "https://example.com/new-avatar.jpg"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Author not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an author",
        "tags": [
          "Authors"
        ]
      },
      "delete": {
        "description": "Deletes an existing author profile. Requires authentication and ownership.",
        "operationId": "AuthorController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the author to delete",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The author has been successfully deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Author not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an author",
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{id}/avatar": {
      "post": {
        "operationId": "AuthorController_uploadAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      },
      "delete": {
        "operationId": "AuthorController_removeAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/avatar": {
      "get": {
        "operationId": "AuthorController_getAvatar",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/cruxes": {
      "get": {
        "operationId": "AuthorController_getPublicCruxes",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/cruxes/{slug}": {
      "get": {
        "operationId": "AuthorController_getCruxBySlug",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/graph": {
      "get": {
        "operationId": "AuthorController_getGraph",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/cruxes/{slug}/artifacts": {
      "get": {
        "operationId": "AuthorController_getPublicArtifacts",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
    "/authors/{identifier}/cruxes/{slug}/artifacts/{artifactId}/download": {
      "get": {
        "operationId": "AuthorController_downloadPublicArtifact",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crux Garden API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the author",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "description": "Username for the author (alphanumeric, hyphens, underscores only)",
            "example": "johndoe",
            "minLength": 1
          },
          "displayName": {
            "type": "string",
            "description": "Display name for the author",
            "example": "John Doe"
          },
          "bio": {
            "type": "string",
            "description": "Bio or description of the author",
            "example": "Software developer and writer"
          },
          "rootId": {
            "type": "string",
            "description": "Root Crux ID for the author",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "homeId": {
            "type": "string",
            "description": "Home ID (set internally from primary home)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "username",
          "displayName"
        ]
      },
      "UpdateAuthorDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username for the author (alphanumeric, hyphens, underscores only)",
            "example": "johndoe"
          },
          "displayName": {
            "type": "string",
            "description": "Display name for the author",
            "example": "John Doe"
          },
          "bio": {
            "type": "string",
            "description": "Bio or description of the author",
            "example": "Software developer and writer"
          },
          "rootId": {
            "type": "string",
            "description": "Root Crux ID for the author",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "meta": {
            "type": "object",
            "description": "Metadata JSON object"
          }
        }
      },
...
    }
  }
}